Add files via upload
help me with this one please. it seems like my carlist is not shared on all the frames or in other words when i login with via admin then i added or removed a car then i press back and login again, it still shows the default cars in displaycars, it seems like theres no car added or removed. same as when i login as admin and add or remove a car then press back and login as user, the check available cars is not updated. how to fix this? help me please